Commit 1.2

+++ src/Commits.java
@@ -1,13 +1,13 @@
 public class Commits {
-    private int[] nums;
+    private double[] nums;
 
 
-    public Commits(int[] nums) {
+    public Commits(double[] nums) {
         this.nums = nums;
     }
 
-    public int getMax() {
-        int rezult = nums[0];
+    public double getMax() {
+        double rezult = nums[0];
 
         for (int i = 1; i < nums.length; i++) {
             if (nums[i] > rezult)
