+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int[] nums = {1, 10, 0, -6};
+        double[] nums = {1, 10, 0, -6};
 
         Commits commits = new Commits(nums);
